// model/planRequest.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* a plan
*/
export class PlanRequest {
    /**
    * The amount in cents that will be charged on the interval specified.
    */
    'amount': number;
    /**
    * ISO 4217 for currencies, for the Mexican peso it is MXN/USD
    */
    'currency'?: string;
    /**
    * Number of repetitions of the frequency NUMBER OF CHARGES TO BE MADE, considering the interval and frequency, this evolves over time, but is subject to the expiration count.
    */
    'expiryCount'?: number;
    /**
    * Frequency of the charge, which together with the interval, can be every 3 weeks, every 4 months, every 2 years, every 5 fortnights
    */
    'frequency': number;
    /**
    * internal reference id
    */
    'id'?: string;
    /**
    * The interval of time between each charge.
    */
    'interval': PlanRequest.IntervalEnum;
    /**
    * The name of the plan.
    */
    'name': string;
    /**
    * The number of days the customer will have a free trial.
    */
    'trialPeriodDays'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "amount",
            "baseName": "amount",
            "type": "number"
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string"
        },
        {
            "name": "expiryCount",
            "baseName": "expiry_count",
            "type": "number"
        },
        {
            "name": "frequency",
            "baseName": "frequency",
            "type": "number"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "interval",
            "baseName": "interval",
            "type": "PlanRequest.IntervalEnum"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "trialPeriodDays",
            "baseName": "trial_period_days",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return PlanRequest.attributeTypeMap;
    }
}

export namespace PlanRequest {
    export enum IntervalEnum {
        Week = <any> 'week',
        HalfMonth = <any> 'half_month',
        Month = <any> 'month',
        Year = <any> 'year'
    }
}