// model/planUpdateRequest.ts
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* a plan
*/
export class PlanUpdateRequest {
    /**
    * The amount in cents that will be charged on the interval specified.
    */
    'amount'?: number;
    /**
    * ISO 4217 for currencies, for the Mexican peso it is MXN/USD
    */
    'currency'?: string;
    /**
    * Number of repetitions of the frequency NUMBER OF CHARGES TO BE MADE, considering the interval and frequency, this evolves over time, but is subject to the expiration count.
    */
    'expiryCount'?: number;
    /**
    * The name of the plan.
    */
    'name'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "amount",
            "baseName": "amount",
            "type": "number"
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string"
        },
        {
            "name": "expiryCount",
            "baseName": "expiry_count",
            "type": "number"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return PlanUpdateRequest.attributeTypeMap;
    }
}