// model/plan-update-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conekta API
 * Conekta sdk
 *
 * The version of the OpenAPI document: 2.1.0
 * Contact: engineering@conekta.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * a plan
 * @export
 * @interface PlanUpdateRequest
 */
export interface PlanUpdateRequest {
    /**
     * The amount in cents that will be charged on the interval specified.
     * @type {number}
     * @memberof PlanUpdateRequest
     */
    'amount'?: number;
    /**
     * ISO 4217 for currencies, for the Mexican peso it is MXN/USD
     * @type {string}
     * @memberof PlanUpdateRequest
     */
    'currency'?: string;
    /**
     * Number of repetitions of the frequency NUMBER OF CHARGES TO BE MADE, considering the interval and frequency, this evolves over time, but is subject to the expiration count.
     * @type {number}
     * @memberof PlanUpdateRequest
     */
    'expiry_count'?: number;
    /**
     * The name of the plan.
     * @type {string}
     * @memberof PlanUpdateRequest
     */
    'name'?: string;
}